requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.30,<3.0.0
alembic>=1.13.2,<2.0.0
aiosqlite>=0.20.0 
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
bcrypt>=4.0.1
PyJWT>=2.8.0

python-multipart>=0.0.9
click>=8.1.7
Jinja2>=3.1.4
ipython>=8.27.0
pytest>=8.3.3
pytest-asyncio>=0.23.8
httpx>=0.27.0
celery[redis]>=5.4.0
asyncpg>=0.31
